Merge dev into main

+++ src/main/java/com/mavi/wishlist/controller/UserController.java
@@ -1,8 +1,10 @@
 package com.mavi.wishlist.controller;
 
 import com.mavi.wishlist.controller.utils.SessionUtils;
+import com.mavi.wishlist.exceptions.InvalidFieldsException;
 import com.mavi.wishlist.model.User;
 import com.mavi.wishlist.service.UserService;
+import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -31,31 +33,20 @@ public class UserController {
     }
 
     @PostMapping("/register")
-    public String addNewUser(HttpSession session, RedirectAttributes redirectAttributes, @ModelAttribute User newUser) {
+    public String addNewUser(HttpSession session, Model model, @ModelAttribute User newUser, HttpServletResponse response) {
 
-        //validate if user with given email already exists
-        if (service.mailIsTaken(newUser.getMail())) {
-            redirectAttributes.addFlashAttribute("showErrorMessage", true);
-            redirectAttributes.addFlashAttribute("errorMessageText", "That email is already in use");
-            System.out.println("controller: failed mailIsTaken check");
-            return "redirect:/user/register";
+        try{
+            newUser = service.registerUser(newUser);
+            session.setAttribute("user", newUser);
+        } catch (InvalidFieldsException e) {
+            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+            model.addAttribute("error", true);
+            model.addAttribute("invalidField", e.getIncorrectField());
+            model.addAttribute("newUser", newUser);
+            return "registerPage";
         }
 
-        System.out.println("controller: passed mailIsTaken check");
-
-        //If user was not added successfully, return value is null, and we redirect to register page with an error message
-        if ((newUser = service.registerUser(newUser)) == null) {
-            redirectAttributes.addFlashAttribute("showErrorMessage", true);
-            redirectAttributes.addFlashAttribute("errorMessageText", "Failed to register user, please try again");
-            System.out.println("controller: registerUser call returned null!");
-            return "redirect:/user/register";
-        }
-
-        System.out.println("controller: passed all checks, redirecting to /wishlist with session attribute");
-
-        session.setAttribute("user", newUser);
-
-        return "redirect:/wishlist/";
+        return "redirect:/wishlist/view";
     }
 
 
@@ -68,26 +59,32 @@ public class UserController {
     }
 
     @PostMapping("/login")
-    public String postLogin(HttpSession session, RedirectAttributes redirectAttributes, @ModelAttribute User user){
+    public String postLogin(HttpSession session, Model model, @ModelAttribute User user, HttpServletResponse response){
 
+        //if credentials are invalid, return to login-form
         if(!service.userLogin(user)) {
-            redirectAttributes.addFlashAttribute("badCredentials", true);
-            return "redirect:/user/login";
+            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+            model.addAttribute("error", true);
+            model.addAttribute("userLogin", user);
+            return "loginPage";
         }
 
-        //update user object with all fields including ID
+        //if credentials are valid, update user object with all fields including ID
         user = service.getUserByMail(user.getMail());
 
+        //add user to session
         session.setAttribute("user", user);
 
-        return "redirect:/wishlist/";
+        return "redirect:/wishlist/view";
     }
 
     @GetMapping("/profile")
     public String showProfile(@RequestParam(required = false, defaultValue = "view") String viewMode, Model model, HttpSession session){
+
         if (!SessionUtils.isLoggedIn(session)) {
             return "redirect:/";
         }
+
         User user = (User) session.getAttribute("user");
         model.addAttribute("user", user);
         model.addAttribute("viewMode", viewMode);
@@ -95,16 +92,21 @@ public class UserController {
     }
 
     @PostMapping("/profile/update")
-    public String updateUser(HttpSession session, RedirectAttributes redirectAttributes, @ModelAttribute User updatedUser){
+    public String updateUser(HttpSession session, RedirectAttributes redirectAttributes, @ModelAttribute User updatedUser, HttpServletResponse response){
+
+        if (!SessionUtils.isLoggedIn(session)) {
+            return "redirect:/";
+        }
 
         updatedUser.setId(this.getUserIdFromSession(session));
 
-        if ((updatedUser = service.updateUser(updatedUser)) == null) {
-            redirectAttributes.addFlashAttribute("showErrorMessage", true);
-            redirectAttributes.addFlashAttribute("errorMessageText", "Failed to update user, please try again");
-            System.out.println("controller: updateUser call returned null!");
-        } else {
+        try{
+            updatedUser = service.updateUser(updatedUser);
             session.setAttribute("user", updatedUser);
+        } catch (InvalidFieldsException e) {
+            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+            redirectAttributes.addFlashAttribute("error", true);
+            redirectAttributes.addFlashAttribute("invalidField", e.getIncorrectField());
         }
 
         return "redirect:/user/profile";
@@ -113,15 +115,26 @@ public class UserController {
     @PostMapping("profile/delete")
     public String deleteUser(HttpSession session, @ModelAttribute User userToDelete) {
 
+        if (!SessionUtils.isLoggedIn(session)) {
+            return "redirect:/";
+        }
+
         userToDelete.setId(this.getUserIdFromSession(session));
 
         this.service.deleteUser(userToDelete);
+
+        session.invalidate();
 
         return "redirect:/";
     }
 
     @PostMapping("profile/logout")
     public String logout(HttpSession session){
+
+        if (!SessionUtils.isLoggedIn(session)) {
+            return "redirect:/";
+        }
+
         session.invalidate();
         return "redirect:/";
     }
